@@ -23,18 +23,18 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/architect": "^0.1802.8",
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.8",
     "@angular/cdk": "17.3.10",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
+    "@angular/common": "^18.2.8",
+    "@angular/compiler": "^18.2.8",
+    "@angular/core": "^18.2.8",
+    "@angular/forms": "^18.2.8",
     "@angular/material": "17.3.10",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/platform-server": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@angular/ssr": "^17.3.10",
+    "@angular/platform-browser": "^18.2.8",
+    "@angular/platform-browser-dynamic": "^18.2.8",
+    "@angular/platform-server": "^18.2.8",
+    "@angular/router": "^18.2.8",
+    "@angular/ssr": "^18.2.8",
     "@angular/youtube-player": "^18.2.8",
     "@apollo/client": "^3.8.3",
     "@apollo/server": "^4.9.3",
@@ -70,14 +70,14 @@
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.10",
+    "@angular-devkit/build-angular": "^18.2.8",
     "@angular-eslint/builder": "15.2.1",
     "@angular-eslint/eslint-plugin": "^16.1.2",
     "@angular-eslint/eslint-plugin-template": "15.2.1",
     "@angular-eslint/schematics": "^18.3.1",
     "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "~17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular/cli": "~18.2.8",
+    "@angular/compiler-cli": "^18.2.8",
     "@capacitor/assets": "^2.0.4",
     "@capacitor/cli": "^5.0.0",
     "@capacitor/splash-screen": "^5.0.0",
