@@ -5,7 +5,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { MaterialModule } from './material.module';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { NgCircleProgressModule } from 'ng-circle-progress';
 import { ContainerComponent } from './components/UI/container/container.component';
@@ -14,28 +14,22 @@ import { RouteReuseStrategy } from '@angular/router';
 import { CustomRouteReuseStrategy } from './utils/CustomRouteReuseStrategy';
 import { MAT_SNACK_BAR_DEFAULT_OPTIONS } from '@angular/material/snack-bar';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-  ],
-  imports: [
-    BrowserModule.withServerTransition({ appId: 'serverApp' }),
-    BrowserAnimationsModule,
-    MaterialModule,
-    AppRoutingModule,
-    HttpClientModule,
-    ContainerComponent,
-    NgCircleProgressModule.forRoot({}),
-    GraphQLModule
-  ],
-  providers: [
-    Title,
-    {
-      provide: RouteReuseStrategy,
-      useClass: CustomRouteReuseStrategy,
-    },
-    { provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: {verticalPosition: 'top'}}
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule.withServerTransition({ appId: 'serverApp' }),
+        BrowserAnimationsModule,
+        MaterialModule,
+        AppRoutingModule,
+        ContainerComponent,
+        NgCircleProgressModule.forRoot({}),
+        GraphQLModule], providers: [
+        Title,
+        {
+            provide: RouteReuseStrategy,
+            useClass: CustomRouteReuseStrategy,
+        },
+        { provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: { verticalPosition: 'top' } },
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule {}
